+++ src/MonadIO/Process.hs
@@ -10,7 +10,7 @@ import Data.Bool               ( otherwise )
 import Data.Ord                ( (>) )
 import Control.Monad           ( join, return )
 import Control.Monad.IO.Class  ( MonadIO, liftIO )
-import Data.Function           ( ($), id )
+import Data.Function           ( ($) )
 import GHC.Stack               ( HasCallStack )
 import System.Exit             ( ExitCode( ExitFailure, ExitSuccess ) )
 
@@ -22,10 +22,6 @@ import Data.Function.Unicode  ( (∘) )
 
 import FPath.Error.FPathError  ( AsFPathError )
 
--- lens --------------------------------
-
-import Control.Lens.Lens  ( lens )
-
 -- monaderror-io -----------------------
 
 import MonadError           ( ѥ )
@@ -34,7 +30,6 @@ import MonadError.IO.Error  ( AsIOError )
 -- more-unicode ------------------------
 
 import Data.MoreUnicode.Either  ( pattern 𝕷, pattern 𝕽 )
-import Data.MoreUnicode.Lens    ( (⊣) )
 import Data.MoreUnicode.Monad   ( (≫) )
 
 -- mtl ---------------------------------
@@ -50,9 +45,7 @@ import System.Process  ( ProcessHandle, waitForProcess )
 ------------------------------------------------------------
 
 import MonadIO.Error.CreateProcError   ( AsCreateProcError )
-import MonadIO.Process.CmdSpec         ( CmdSpec, HasCmdSpec( cmdSpec ) )
-import MonadIO.Process.CreateProcOpts  ( CreateProcOpts
-                                       , HasCreateProcOpts( createProcOpts ) )
+import MonadIO.Process.CmdSpec         ( CmdSpec )
 import MonadIO.Process.ExitStatus      ( ExitStatus( ExitSig, ExitVal ) )
 import MonadIO.Process.MakeProc        ( MakeProc, makeProc )
 import MonadIO.Process.MkStream        ( MkStream )
@@ -60,16 +53,6 @@ import MonadIO.Process.Signal          ( Signal( Signal ) )
 import MonadIO.Process.OutputHandles   ( OutputHandles( slurp ) )
 
 --------------------------------------------------------------------------------
-
-data Cmd = Cmd CmdSpec CreateProcOpts
-
-instance HasCmdSpec Cmd where
-  cmdSpec = lens (\ (Cmd spec _) → spec) (\ (Cmd _ cpo) spec → Cmd spec cpo)
-
-instance HasCreateProcOpts Cmd where
-  createProcOpts = lens (\ (Cmd _ cpo) → cpo) (\ (Cmd spec _) cpo→ Cmd spec cpo)
-
-------------------------------------------------------------
 
 exitCode ∷ ExitCode → ExitStatus
 exitCode ExitSuccess     = ExitVal 0
@@ -92,17 +75,18 @@ procWait prox = do
 ----------------------------------------
 
 system ∷ (MonadIO μ, HasCallStack, MkStream σ,
-          MakeProc ζ, OutputHandles ζ ω, HasCreateProcOpts φ,
+          MakeProc ζ, OutputHandles ζ ω,
           AsCreateProcError ε, AsFPathError ε, AsIOError ε, MonadError ε μ) ⇒
-         σ → (φ,CmdSpec) → μ (ExitStatus, ω)
+         σ → CmdSpec → μ (ExitStatus, ω)
 
-system inh (opts,cspec) = do
-  x ← ѥ $ makeProc (opts ⊣ createProcOpts) inh cspec
+system inh cspec = do
+  x ← ѥ $ makeProc inh cspec
   ѥ x ≫ \case
     𝕷 e → join ∘ return $ throwError e
     𝕽 r → procWait (return r)
 
 -- $ system defCPOpts (""∷ Text) (CmdSpec (CmdExe [absfile|/usr/bin/env|]) (CmdArgs []))
 
+-- splitMError @ProcError @(Either _) $ system ("" :: Text) (mkCmd [absfile|/usr/bin/env|]  [])
 
 -- that's all, folks! ----------------------------------------------------------